(#1114) Split store_datacollection_in_ispyb into separate modules and rename it to xxx_ispyb_store
