+++ tests/system_tests/external_interaction/test_ispyb_dev_connection.py
@@ -12,11 +12,17 @@ from hyperion.experiment_plans.rotation_scan_plan import (
 from hyperion.external_interaction.callbacks.rotation.callback_collection import (
     RotationCallbackCollection,
 )
-from hyperion.external_interaction.ispyb.store_datacollection_in_ispyb import (
-    IspybIds,
-    Store2DGridscanInIspyb,
-    Store3DGridscanInIspyb,
+from hyperion.external_interaction.ispyb.gridscan_ispyb_store import (
     StoreGridscanInIspyb,
+)
+from hyperion.external_interaction.ispyb.gridscan_ispyb_store_2d import (
+    Store2DGridscanInIspyb,
+)
+from hyperion.external_interaction.ispyb.gridscan_ispyb_store_3d import (
+    Store3DGridscanInIspyb,
+)
+from hyperion.external_interaction.ispyb.ispyb_store import (
+    IspybIds,
 )
 from hyperion.parameters.constants import DEV_ISPYB_DATABASE_CFG
 from hyperion.parameters.external_parameters import from_file as default_raw_params
